Fixed audio stream renderer; audio data is now dynamically loaded into the heap instead of being allocated in statically reserved memory, allowing larger files to be played; added more logs; README.md updated

+++ audio_format_readers/wav/wav_reader.hpp
@@ -11,7 +11,7 @@ class WAVReader {
 private:
     char file_path[MAX_FILE_PATH_SIZE];
     uint32_t audio_buffer_size;
-    float *audio_buffer;
+    BYTE *audio_buffer;
 public:
     WAVReader();
